refactor: Removing code connected to old migrations
