+++ zmessaging/src/main/scala/com/waz/content/AccountsStorage.scala
@@ -19,13 +19,9 @@ package com.waz.content
 
 import android.content.Context
 import com.waz.model.AccountData.AccountDataDao
-import com.waz.model.AccountDataOld.AccountDataOldDao
 import com.waz.model._
 import com.waz.utils.TrimmingLruCache.Fixed
 import com.waz.utils.{CachedStorage, CachedStorageImpl, TrimmingLruCache}
 
 trait AccountStorage extends CachedStorage[UserId, AccountData]
 class AccountStorageImpl(context: Context, storage: Database) extends CachedStorageImpl[UserId, AccountData](new TrimmingLruCache(context, Fixed(8)), storage)(AccountDataDao) with AccountStorage
-
-trait AccountsStorageOld extends CachedStorage[AccountId, AccountDataOld]
-class AccountsStorageOldImpl(context: Context, storage: Database) extends CachedStorageImpl[AccountId, AccountDataOld](new TrimmingLruCache(context, Fixed(8)), storage)(AccountDataOldDao) with AccountsStorageOld
